created menage list when user is logged in

+++ routes/web.php
@@ -23,26 +23,30 @@ Route::get('/',[ListingController::class,'index']);
 
 
 
-Route::get('/listings/create',[ListingController::class,'create']);
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
 
 Route::post('/listings',[ListingController::class,'store']);
 
 Route::get('/listings/{listing}/edit',
-[ListingController::class,'edit']);
+[ListingController::class,'edit'])->middleware('auth');
 
-Route::put('/listings/{listing}',[ListingController::class,'update']);
+Route::put('/listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
-Route::get('/register',[UserController::class,'create']);
+Route::get('/listings/menage',[ListingController::class,'menage'])->middleware('auth');
+
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 
 Route::post('/users',[UserController::class,'store']);
 
 Route::post('/logout',[UserController::class,'logout']);
 
-Route::get('/login',[UserController::class,'login']);
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
 
 Route::post('/users/authenticate',[UserController::class,'authenticate']);
+
+
